requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-environ==0.11.2
django-import-export==3.3.7
django-unfold==0.21.1
et-xmlfile==1.1.0
html2image==2.0.4.3
idna==3.7
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.1
pdfkit==1.0.0
pillow==10.3.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.7.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
